requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
xgboost==1.7.6
lightgbm==3.3.5
imbalanced-learn==0.10.1
shap==0.41.0
mlflow==2.4.0
flask==2.3.2
streamlit==1.20.0
Pillow==9.5.0
requests==2.31.0
plotly==5.14.1
evidently==0.4.40
fastapi~=0.98.0
uvicorn~=0.22.0
